# improve performance of RectHelper.Extend