```

BenchmarkDotNet v0.13.10, Windows 10 (10.0.19044.3693/21H2/November2021Update)
.NET SDK 8.0.100
  [Host]     : .NET 7.0.14 (7.0.1423.51910), X64 RyuJIT AVX2
  Job-LZKMVF : .NET 7.0.14 (7.0.1423.51910), X64 RyuJIT AVX2
  Job-NCDEPJ : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2

IterationCount=10  LaunchCount=1  WarmupCount=1  

```
| Method     | Runtime  | Mean     | Error   | StdDev  | Ratio | Allocated | Alloc Ratio |
|----------- |--------- |---------:|--------:|--------:|------:|----------:|------------:|
| ExtendBySdcb | .NET 7.0 | 772.9 ns | 7.73 ns | 5.12 ns |  1.00 |         - |          NA |
| ExtendByAven | .NET 7.0 | 458.9 ns | 5.23 ns | 2.73 ns |  0.59 |         - |          NA |
|            |          |          |         |         |       |           |             |
| ExtendBySdcb | .NET 8.0 | 227.0 ns | 3.26 ns | 2.16 ns |  1.00 |         - |          NA |
| ExtendByAven | .NET 8.0 | 176.6 ns | 0.92 ns | 0.55 ns |  0.77 |         - |          NA |

the code of benchmark.

```
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Jobs;
using BenchmarkDotNet.Running;
using OpenCvSharp;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Sdcb.OpenVINO.Benchmarks
{
    [SimpleJob(runtimeMoniker: RuntimeMoniker.Net70, launchCount: 1, warmupCount: 1, iterationCount: 10)]
    [SimpleJob(runtimeMoniker: RuntimeMoniker.Net80, launchCount: 1, warmupCount: 1, iterationCount: 10)]
    [MemoryDiagnoser]
    [HtmlExporter]
    [MarkdownExporterAttribute.GitHub]
    public class RectExtend
    {
        private Random random;
        private List<Rect> rects;

        [GlobalSetup]
        public void GlobalSetup()
        {
            this.random = new();
            int rectNum = 100;
            this.rects = new(rectNum);
            for (int i = 0; i < rectNum; i++)
            {
                this.rects.Add(new Rect(
                    this.random.Next(0, 640), this.random.Next(0, 640),
                    this.random.Next(0, 640), this.random.Next(0, 640))
                    );
            }
        }

        [Benchmark(Description = "ExtendBySdcb", Baseline = true)]
        public void ExtendBySdcb()
        {
            for (int i = 0; i < this.rects.Count; i += 2)
            {
                ExtendSdcb(this.rects[i], this.random.Next(0, 88));
            }
        }

        [Benchmark(Description = "ExtendByAven")]
        public void ExtendByAven()
        {
            for (int i = 0; i < this.rects.Count; i += 2)
            {
                ExtendAven(this.rects[i], this.random.Next(0, 88));
            }
        }

        static Rect ExtendSdcb(in Rect rect, int extendLength)
        {
            return Rect.FromLTRB(rect.Left - extendLength, rect.Top - extendLength, rect.Right + extendLength, rect.Bottom + extendLength);
        }

        static Rect ExtendAven(in Rect rect, int extendLength)
        {
            rect.Inflate(extendLength, extendLength);
            return rect;
        }
    }
}