@sdcb 2023-12-01
The problem of this is imcoming Rect should be const, this would impact the imcoming Rect, I think?