@AvenSun 2023-12-01
sorry for that, the keyword **in** is ignored.
fixed now. new benchmark result as fellows.

```

BenchmarkDotNet v0.13.10, Windows 10 (10.0.19044.3693/21H2/November2021Update)
.NET SDK 8.0.100
  [Host]     : .NET 7.0.14 (7.0.1423.51910), X64 RyuJIT AVX2
  Job-PXWREW : .NET 7.0.14 (7.0.1423.51910), X64 RyuJIT AVX2
  Job-LCQJJH : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2

IterationCount=10  LaunchCount=1  WarmupCount=1  

```
| Method       | Runtime  | Mean       | Error    | StdDev  | Ratio | Allocated | Alloc Ratio |
|------------- |--------- |-----------:|---------:|--------:|------:|----------:|------------:|
| ExtendBySdcb | .NET 7.0 | 1,477.7 ns | 10.54 ns | 6.27 ns |  1.00 |         - |          NA |
| ExtendByAven | .NET 7.0 | 1,031.4 ns |  6.97 ns | 4.61 ns |  0.70 |         - |          NA |
|              |          |            |          |         |       |           |             |
| ExtendBySdcb | .NET 8.0 |   441.5 ns |  1.91 ns | 1.26 ns |  1.00 |         - |          NA |
| ExtendByAven | .NET 8.0 |   325.4 ns |  0.99 ns | 0.52 ns |  0.74 |         - |          NA |