@AvenSun 2023-12-05
hi @sdcb 
sorry for the delay.
benchmark without in keyword as follows

```

BenchmarkDotNet v0.13.10, Windows 10 (10.0.19044.3693/21H2/November2021Update)
.NET SDK 8.0.100
  [Host]     : .NET 7.0.14 (7.0.1423.51910), X64 RyuJIT AVX2
  Job-ERUGJG : .NET 7.0.14 (7.0.1423.51910), X64 RyuJIT AVX2
  Job-HGGJRE : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2

IterationCount=10  LaunchCount=1  WarmupCount=1  

```
| Method       | Runtime  | Mean       | Error     | StdDev    | Ratio | RatioSD | Allocated | Alloc Ratio |
|------------- |--------- |-----------:|----------:|----------:|------:|--------:|----------:|------------:|
| ExtendBySdcb | .NET 7.0 | 1,939.1 ns | 480.54 ns | 317.85 ns |  1.00 |    0.00 |         - |          NA |
| ExtendByAven | .NET 7.0 |   995.2 ns |  39.29 ns |  25.99 ns |  0.53 |    0.09 |         - |          NA |
|              |          |            |           |           |       |         |           |             |
| ExtendBySdcb | .NET 8.0 |   440.6 ns |   0.77 ns |   0.46 ns |  1.00 |    0.00 |         - |          NA |
| ExtendByAven | .NET 8.0 |   326.1 ns |   1.07 ns |   0.64 ns |  0.74 |    0.00 |         - |          NA |